Reimplement printing the best solution to enchance code reuse
The function for printing the board can be used to print the
best solution. Doing this can prevent code duplication and
inconsistent implementations for the same porpose.

Format people often use when writin commit messages:
One-line summary of 50-100 characters,
followed by an empty line,
then some detailed explanations about the patch, if needed.

We do not refer to code, but focus on describing why the
changes have been made and how.

Present tense is used.